// TrixxDiscordBot/trixxdiscordbot.client/src/api/models/auth-result-model.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { Permission } from '../models/permission';
export interface AuthResultModel {
  accessToken?: string | null;
  error?: string | null;
  permissions?: Array<Permission> | null;
  refreshToken?: string | null;
  success: boolean;
}